tests: test multiple feature combinations
