+++ pyproject.toml
@@ -22,12 +22,11 @@ classifiers = [
 ]
 dynamic = [ "version" ]
 dependencies = [
-  "anndata",
   # for debug logging (referenced from the issue template)
   "session-info2",
   "typing-extensions; python_version<'3.13'",
 ]
-optional-dependencies.datasets = [ "pooch", "pyyaml", "tqdm" ]
+optional-dependencies.datasets = [ "anndata", "pooch", "pyyaml", "tqdm" ]
 optional-dependencies.settings = [ "pydantic-settings", "python-dotenv" ]
 optional-dependencies.spatialdata = [ "spatialdata" ]
 optional-dependencies.sphinx = [ "pydocstring-rs>=0.1.13", "sphinx>=9" ]
@@ -41,14 +40,7 @@ dev = [
   "pre-commit",
   "twine>=4.0.2",
 ]
-test = [
-  "coverage>=7.10",
-  "numpy",
-  "pytest",
-  "scverse-misc[datasets,settings,sphinx]",
-  "sphinx",
-  "sphinx-autodoc-typehints"
-]
+test = [ "coverage>=7.10", "numpy", "pytest", "sphinx", "sphinx-autodoc-typehints" ]
 doc = [
   "ipykernel",
   "ipython",
@@ -77,6 +69,8 @@ envs.docs.dependency-groups = [ "doc" ]
 envs.hatch-test.matrix = [
   # Test the lowest and highest supported Python versions with normal deps
   { deps = [ "stable" ], python = [ "3.12", "3.13", "3.14" ] },
+  # Test the newest supported Python version with different feature sets
+  { deps = [ "stable" ], python = [ "3.14" ], features = [ "min", "sphinx", "full" ] },
   # Test the newest supported Python version also with pre-release deps
   { deps = [ "pre" ], python = [ "3.14" ] },
 ]
@@ -84,6 +78,13 @@ envs.hatch-test.matrix = [
 # set the environment variable `UV_PRERELEASE` to "allow".
 envs.hatch-test.overrides.matrix.deps.env-vars = [
   { value = "allow", key = "UV_PRERELEASE", if = [ "pre" ] },
+]
+envs.hatch-test.overrides.matrix.deps.features = [
+  # make sure the sphinx extension works without having all deps
+  { value = "sphinx", if = [ "sphinx", "full" ] },
+  { value = "datasets", if = [ "full" ] },
+  { value = "spatialdata", if = [ "full" ] },
+  { value = "settings", if = [ "full" ] },
 ]
 envs.hatch-test.dependency-groups = [ "dev", "test" ]
