+++ tests/settings/conftest.py
@@ -1,0 +1,3 @@
+import pytest
+
+pytest.importorskip("scverse_misc.settings")
